[라테] 로또 미션 2단계 제출합니다. (#170)
* [refactor] stream 구조 수정 및 메소드 위치 수정

ResultCount 불필요한 형변환 제거
LottoGame 메소드 위치 변경 및 stream 수정
ResultCountTest 불필요한 공백 제거

* [feat] RepeatCount 클래스 추가

RepeatCount 로또 게임의 반복 횟수를 담은 객체
RepeatCountTest RepeatCount 테스트

* [refactor] Money 클래스가 RepeatCount를 반환하도록 수정

* [feat] LottoGame merge 기능 추가

LottoGame 다음 수동 구매 기능 추가를 위해 자동구매와 수동구매 게임을 합치는 merge기능 추가
LottoGameTest merge 테스트

* [feat] 로또 수동구매 기능 추가

UserNumberGenerator UserInput을 받아 직접 번호를 입력해주는 클래스
InputView User Number와 Numbers를 받을 때 메세지를 인자로 넣어줘 반복되는 기능 삭제
OutputView 메세지 추가
LottoController 기능 연결을 위한 코드 수정

* [feat] 번호 입력 메세지를 한 번만 출력하기 위한 클래스 기능 추가

LottoController RepeatCount를 받아 메세지 출력 기능 구현
UserNumberGenerator 메세지 출력 기능 삭제
RepeatCount isNotZero를 통해 수동 입력이 1 이상일 경우를 알려줌
OutputView RepeatCount가 0이 아닐 경우 입력 메세지 출력

* [refactor] LottoNumberFactory 생성자 private으로 변경

* [refactor] generator 패키지 위치 변경 및 기능 수정

*generator 모든 generator가 domain과의 의존성을 없애기 위해 Integer List를 반환하도록 조정

* [refactor] 불필요한 import 삭제 및 메소드 순서 및 구현 수정, 오타 삭제

* [refactor] OutputView 메소드 중 printRequestUserNumbers 인자를 boolean형을 직접 받도록 수정

* [refactor] 기타 컨벤션 수정

* [docs] 로또 수동기능 구현에 따른 README.md 수정

* [refactor] test 코드 전체적인 refactor

All DisplayName을 이용해 경고 없애기, 테스트 이름을 보다 구체적으로 수정하기

* [refactor] LottoWinner를 int와 List<Integer>로 직접 생성할 수 있도록 수정

LottoWinner create 메소드 추가 및 생성자 private로 변경
LottoGame 생성자 private로 변경

+++ src/main/java/domain/lotto/lottoresult/ResultCount.java
@@ -10,7 +10,7 @@ public class ResultCount {
     }
 
     public long multiply(int operand) {
-        return (long) count * operand;
+        return count * operand;
     }
 
     public static ResultCount sum(ResultCount count1, ResultCount count2) {
